Needless optimisation of day1

+++ day01/src/part1.zig
@@ -1,8 +1,8 @@
 const std = @import("std");
 const day1 = @import("root.zig");
 
-fn part1(left: *std.ArrayList(i64), right: *std.ArrayList(i64)) i128 {
-    var sum: i128 = 0;
+fn part1(left: *std.ArrayList(i32), right: *std.ArrayList(i32)) i32 {
+    var sum: i32 = 0;
     for (left.items, right.items) |l, r| {
         sum += @max(l, r) - @min(l, r);
     }
